requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
alembic==1.13.2
jinja2==3.1.4
loguru==0.7.2
python-multipart==0.0.9
email-validator==1.3.1
passlib[bcrypt]
PyJWT